メンション送信部分を共通
